+++ src/Components/Button/Button.tsx
@@ -1,4 +1,3 @@
-import classNames from "classnames";
 import React from "react";
 import { FunctionComponent } from "react";
