+++ src/Components/Loader/Loader.tsx
@@ -1,0 +1,12 @@
+import "./loader.css";
+
+export const Loader = () => {
+  return (
+    <div id="dots1">
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+    </div>
+  );
+};
